requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.40
langchain-openai>=0.0.8
langchain-community>=0.0.20
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
python-pptx>=0.6.21
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0

# Development and testing dependencies
pytest>=6.0
pytest-cov>=2.0
black>=21.0
flake8>=3.8
mypy>=0.800
pre-commit>=2.15
